agent/requirements.txt
fastapi
uvicorn
python-dotenv
livekit-api
livekit-agents
pinecone
langchain
langchain-openai
langchain-pinecone
openai